Added lesson_02 Fix task_3

+++ lesson_02/lesson_2_task_3.py
@@ -5,6 +5,6 @@ def square(a):
     return math.ceil(a ** 2)
 
 
-side = int(input("Введите сторону: "))
+side = float(input("Введите сторону: "))
 result = square(side)
 print(f"Площадь квадрата со сторой {side}: {result}")
